tracecap_viewer/requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
pymysql
nanoid
PyGithub
boto3
